+++ app/src/main/java/lavender/client/android/data/grpc/GrpcClient.kt
@@ -96,4 +96,8 @@ class GrpcClient {
     fun getAvatarCache(): Map<String, String> {
         return realGrpcClient.getAvatarCache()
     }
+
+    fun updateAvatarCache(username: String, avatarUrl: String) {
+        realGrpcClient.updateAvatarCache(username, avatarUrl)
+    }
 }